# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Codessa API",
    "version": "1.0.0",
    "description": "AI-powered GitHub pull request review system. Authentication uses an httpOnly `session` cookie set after completing the GitHub OAuth login flow — browser requests automatically include it; non-browser clients must forward the `Set-Cookie` value."
  },
  "servers": [
    { "url": "https://api.codessa.xyz", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Set automatically after completing GET /auth/github login flow"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "example": { "error": "Unauthorized" }
      },
...
      "RepositoryConfig": {
        "type": "object",
        "properties": {
          "customInstructions": { "type": "string", "nullable": true, "example": "Always suggest a code snippet for fixes." },
          "tone": { "type": "string", "nullable": true, "enum": ["friendly", "strict", "concise"] },
          "severityThreshold": { "type": "string", "nullable": true, "enum": ["balanced", "critical_only", "strict"] },
          "analysisFocus": {
            "type": "object",
            "nullable": true,
            "properties": {
              "security": { "type": "boolean" },
              "performance": { "type": "boolean" },
              "bugs": { "type": "boolean" },
              "codeStyle": { "type": "boolean" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/repositories": {
      "get": {
        "tags": ["Repositories"],
        "summary": "List repositories accessible via your GitHub App installations",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Case-insensitive substring match on full repo name" }
        ],
        "responses": {
          "200": {
            "description": "Paginated repository list",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "githubRepoId": "123456789",
                      "fullName": "octocat/hello-world",
                      "owner": "octocat",
                      "name": "hello-world",
                      "private": false,
                      "defaultBranch": "main",
                      "installationId": 54321,
                      "enabled": true,
                      "hasCustomConfig": false
                    }
                  ],
                  "page": 1,
                  "perPage": 20,
                  "total": 1,
                  "totalPages": 1
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/repositories/{owner}/{repo}/enable": {
      "post": {
        "tags": ["Repositories"],
        "summary": "Enable auto-review for a repository",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" }, "example": "octocat" },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hello-world" }
        ],
        "responses": {
          "200": { "description": "Enabled", "content": { "application/json": { "example": { "ok": true } } } },
          "404": {
            "description": "Repository not found or Codessa app not installed on it",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/repositories/{owner}/{repo}/disable": {
      "post": {
        "tags": ["Repositories"],
        "summary": "Disable auto-review for a repository",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" }, "example": "octocat" },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hello-world" }
        ],
        "responses": {
          "200": { "description": "Disabled", "content": { "application/json": { "example": { "ok": true } } } }
        }
      }
    },
    "/repositories/{owner}/{repo}/config": {
      "get": {
        "tags": ["Repositories"],
        "summary": "Get dashboard review config for a repository",
        "description": "These are fallback settings used when a field isn't set in the repo's .github/.codessa.yml — the YAML file always wins per-field when present.",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" }, "example": "octocat" },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hello-world" }
        ],
        "responses": {
          "200": {
            "description": "Current dashboard config",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepositoryConfig" } } }
          },
          "404": {
            "description": "Repository not enabled yet",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "patch": {
        "tags": ["Repositories"],
        "summary": "Update dashboard review config for a repository",
        "description": "Only the provided fields are updated. The repository must already be enabled (via /repositories/{owner}/{repo}/enable) before its config can be set.",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" }, "example": "octocat" },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hello-world" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RepositoryConfig" },
              "example": {
                "customInstructions": "Always suggest a code snippet for fixes.",
                "tone": "friendly",
                "severityThreshold": "balanced",
                "analysisFocus": { "security": true, "performance": true, "bugs": true, "codeStyle": false }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated config",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepositoryConfig" } } }
          },
          "400": {
            "description": "Invalid config payload",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Repository not enabled yet",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  }
}